create heading component
