+++ src/core/App.js
@@ -1,18 +1,16 @@
 import React, { Component } from 'react'
-import { fetchCaptionEndpoint } from '../api/cloudmersive.js'
+import { fetchCaptionEndpoint } from '../common/api/cloudmersive.js'
 
 import styles from './App.module.scss'
 
 // Custom Components
-// Heading
-// AppHeader
-// TextWindow
+import { Heading } from '../common/components/Heading/Heading'
 import ImageUploader from '../common/components/ImageUploader/ImageUploader'
-// ConfidenceRater
-// CaptionBox
-// ResetButton
-// CopyButton
-// ProgressBar
+// ConfidenceRater (from the API - confidence % of AI & the caption it generated)
+// CaptionBox (caption input text field)
+// ResetButton (reset uploaded file and generated captions)
+// CopyButton (to copy caption to clipboard)
+// ProgressBar (when caption is being generated)
 
 class App extends Component {
   constructor () {
@@ -59,13 +57,18 @@ class App extends Component {
     return (
       <main className={styles.App}>
         <header>
-          <h1>Image Caption App</h1>
+          <Heading level={1}>Image Caption App</Heading>
         </header>
 
-        <ImageUploader
-          getImageFile={this.getImageFile}
-          getFormData={this.getFormData}
-        />
+        <div className='grid'>
+          <div className='column'>
+            <ImageUploader
+              getImageFile={this.getImageFile}
+              getFormData={this.getFormData}
+            />
+          </div>
+          <div className='column'></div>
+        </div>
 
         <footer>
           <span>Created by M.K. Reyes | &copy; 2019</span>
